third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/duplicate-directive-expected.txt

CONSOLE ERROR: Ignoring duplicate Content-Security-Policy directive 'script-src'.
ALERT: PASS (1/1)
This tests the effect of duplicated directives. It passes if a console warning is visible, and the alert() is executed.
